+++ VisilabsIOS/Classes/TargetingAction/InAppNotification/Models/UIView+Animations.swift
@@ -32,25 +32,4 @@ internal extension UIView {
             self.layoutIfNeeded()
         }, completion: nil)
     }
-    
-    //TODO: kaldır sonra bunu gerek yok.
-    func pv_shake() {
-        layer.removeAnimation(forKey: shakeKey)
-        let vals: [Double] = [-2, 2, -2, 2, 0]
-        
-        let translation = CAKeyframeAnimation(keyPath: "transform.translation.x")
-        translation.timingFunction = CAMediaTimingFunction(name: CAMediaTimingFunctionName.linear)
-        translation.values = vals
-        
-        let rotation = CAKeyframeAnimation(keyPath: "transform.rotation.z")
-        rotation.values = vals.map { (degrees: Double) in
-            let radians: Double = (Double.pi * degrees) / 180.0
-            return radians
-        }
-        
-        let shakeGroup: CAAnimationGroup = CAAnimationGroup()
-        shakeGroup.animations = [translation, rotation]
-        shakeGroup.duration = 0.3
-        self.layer.add(shakeGroup, forKey: shakeKey)
-    }
 }
